Update tests, add circleCI
