+++ circle.yml
@@ -1,0 +1,3 @@
+machine:
+  node:
+    version: 6
